term/requirements.txt
pokereval
retrace==2.2.6
elasticsearch_dsl>=2.0.0,<3.0.0
treelib==1.3.5
sortedcontainers==1.5.4
pillow==3.4.2
numpy==1.11.2
ruamel.yaml==0.13.7
pyvbox==1.0.0
pytesseract==0.1.6
